run_all: auto generate cam_to_isp_config
Summary:
There is no `cam_to_isp_config.json` by default, since it needs absolute paths on it.

This diff checks if it exists, and if not it creates one, and it duplicates the default ISP files to `isp{0..16}.json`. The last part was already done before when isp0.json was not found, but we now do everything at once before populating the fields in run_all.

Reviewed By: fbriggs

Differential Revision: D3838115

fbshipit-source-id: 0e8cc0734b3c8a7964353f31d0cdbd9b392c149e