Color calibration - filter detected contours
Summary:
There were some constraints on the chessboard detector that were causing false positives, and also missing some real patches due to closeness to the border of the image, and their size when the camera is closer.

This diff removes the constraint of ignoring the border of the image, relaxes the max area and reduces the min distance allowed between patches. It also slightly increases the block size of the adaptive threshold algorithm.

Reviewed By: tox-fb

Differential Revision: D3841772

fbshipit-source-id: 0174b390292a5244d778f39d11779155ed0dc451