Color calibration - Relax black level mask
Summary:
To avoid getting false contour positives around the SpyderCube we can move it further away from the checkerboard, but the black level detector mask zeroes out 50% of the image from the center.

This diff relaxes the black level mask to 90% instead. In any case the idea of the mask was to avoid embedded image on the borders and other junk outside the light box.

Reviewed By: fbriggs

Differential Revision: D3845532

fbshipit-source-id: 96011be589dddba792d2624f3b434e0f56659c90