run all - fix automatic generation of cam_to_isp mappings
Summary:
With D3838115 we now automatically generate the cam_to_isp mappings file when this or isp0.json don't exist, but the "or" clause was not correctly formed. So, if we remove `cam_to_isp_config.json` and run `run_all`, it is not re-generated.

After giving it some more thought, I believe the right thing to do is to just check for the existence of `cam_to_isp_config.json`, and create it and `isp{0..16}.json` if it doesn't exist. If we then remove `isp{0..16}.json` we assume we know what we're doing, and we should modify the paths in `cam_to_isp_config.json` to point to the new ISP config files.

Reviewed By: tox-fb

Differential Revision: D3863931

fbshipit-source-id: bec8b0f02756ba0db7eaa20150e0a6066a721559