better algorithm to compute initial flow, perhaps
Summary:
at the coarsest level of the flow pyramid, the algorithm is seeded with
zero flow everywhere. instead, perform a search in the direction specified
by the hint to see if there is a better fit there. the default is to search
zero percent of the smaller image dimension, effectively disabled

it is also unlikely that the error metric is particularly good. running an
off-the-shelf stereo matching algorithm would be interesting

here is a comparison of pixflow_low and pixflow_low_search_low for a case where someone is standing only 2 feet from the camera
{F62835722}
{F62835723}

or compare full eqr frames, before:
{F62916292}
after:
{F62916622}

Reviewed By: fbriggs

Differential Revision: D3719266

fbshipit-source-id: d68ac607a0b4e9c1703bcaa10dfe857e7e0a7ebd