Make multi-camera black level detection more robust
Summary:
In D4140603 (Noise-robust black level computation) we assumed the black level is the lowest non-zero value with enough pixel count (to avoid noise and dead pixels). This is not always true, specially when working with12-bit images. The pixel values in a black hole are more granular, so we need to to a median.

To make it more robust, we first use the histogram approach to find a black level threshold. This gives us a mask of pixels that are candidates to black level. We discard outliers assuming that the black hole is circular, and then we pick lowest median values of all remaining candidates.

Also, we add a new feature in the color calibration code, so that we can align the black level across multiple cameras to their per-channel median.

Minor changes:
- After D4160400 (Read 12-bit images properly using CV_LOAD_IMAGE_ANYDEPTH) we have to make sure we are loading the right images (OpenCV doesn't allow 16-bit images in some of its functions)
- `manual_black_level` now expects per-channel values in the range [0..1] so as to accommodate any bit depth
- Slightly relaxed constraint on centered charts

Reviewed By: fbriggs

Differential Revision: D4160627

fbshipit-source-id: 018cbf0dfb6191bb0ed03d41a51b70a009ec93e5