Add histogram display and fix a race condition that sometimes occurs
Summary:
This diff adds a histogram overlay on top of preview,
as well as fixes a race condition that sometimes occurs when
a user switches bit depth of the camera while preview is running.
This causes USB packet descriptors to be unmapped from memory and
pointers get invalidated at random points in time while various parts
of the program are using the frame data.

Reviewed By: aparrapo

Differential Revision: D4179334

fbshipit-source-id: 22a9a5ee9ee68585e7e6eb5c39016a0c206684a2