Compared various results and put gradient lerp back in.
Summary: I carefully compared the c++ and Halide versions and they are doing the right thing so I don't see a lurking bug per se. I also believe the lerp version looks better with real data so I'm putting that back the way it was. I am seeing a fair bit of chroma ringing and I need to see if that is in the optics or the demosaicer.  But that will be handled in a different diff no matter the source.

Reviewed By: tox-fb

Differential Revision: D4210708

fbshipit-source-id: 98dcb51224850157308a94a879f435051520107d