Unpack without cameranames.txt. Storing and reading metadata on binaries
Summary:
Currently the capturing software saves .bin files and a cameranames.txt file every time. This txt file tells us in what order the frames were written, so if we give the image width and height, and the bit depth to the unpacker it knows how to extract the data.

After D4094948 (Fix an out of bounds access, drive by whitespace / indentation fixes) we had the serial numbers "burned" as metadata on each frame, but we still needed cameranames.txt to know the camera ordering.

This diff adds new functionality to the capturing software so we just create the .bin files, and the unpacker so we don't need cameranames.txt. Now when capturing we use the first 4096 bytes on each .bin file to store metadata info:

1) magic number (tells us if file has metadata)
2) timestamp
3) bin file count
4) total number of bin files
5) image width
6) image height
7) bits per pixel
8) number of unique cameras in bin file

Each frame also has its serial number.

These parameters are enough to unpack everything without needing any extra files or input.

Note that the code is still backwards compatible with the old capturing software. If the directory containing the .bin files does not have the cameranames.txt file we assume all the metadata is on the binaries. The unpacker checks for validity.

Reviewed By: fbriggs

Differential Revision: D4174825

fbshipit-source-id: b67d59dafcee5fe2c7bf1a26bf7ed9c5a27ab7a3