ISP clamping in Halide
Summary:
The color correction method introduced not long ago used two new ISP fields: `clampMin` and `clampMax`. They are used in an additional step called `clampAndStretch` in `CameraIsp.h`, but not in the Halide version.

This diff adds this feature to the Halide version, as part of the `demosaic` step.

From the original equation on the non-Halide version:

  (((x - minRaw) * wb) / (maxRaw - minRaw) - clampMin) / (clampMax - clampMin)
  = ((x - minRaw) * wb - clampMin * (maxRaw - minRaw)) / ((maxRaw - minRaw) * (clampMax - clampMin))

if we define
  rawRange = maxRaw - minRaw

we get
  ((x - minRaw) * wb - clampMin * (maxRaw - minRaw)) / ((maxRaw - minRaw) * (clampMax - clampMin))
  = ((x - minRaw) * wb - clampMin * rawRange) / (rawRange * (clampMax - clampMin))
  = (x * wb - (minRaw * wb + clampMin * rawRange)) / (rawRange * (clampMax - clampMin))
  = wb * (x - (minRaw + clampMin * rawRange / wb)) / (rawRange * (clampMax - clampMin))
  = (x - A) * B / C

where
  A = minRaw + clampMin * rawRange / wb
  B = wb
  C = rawRange * (clampMax - clampMin)

Reviewed By: tox-fb

Differential Revision: D4217944

fbshipit-source-id: 59c0d4615e30bff6c806790b069a45bdcaa7b13b