Unpacker speedup
Summary:
In OpenCV, `imwrite` accepts a third argument, which according to the documentation can contain "format-specific save parameters encoded as pairs." They show examples for JPEG, WEBP, PNG, PPM, PGM and even PBM, but nothing on TIFF.

After a little digging it seems that paramId=259 sets the TIFF data compression technique, and a value of 1 sets it to "dump mode".

I've ran some tests with the unpacker and using that mode we can write to disk ~4 times faster.
The only drawback is that all the images are WxHxnbits/8 bytes, regardless of the image content (as expected).

Reviewed By: superdog8

Differential Revision: D4300822

fbshipit-source-id: 846799a717fda847625216a6120c2b4e963f8bb9