Make chart area thresholds visible to user/script
Summary:
When taking pictures of color charts for color calibration we may get too far/close, use different types of charts, etc. so I think it's a good idea that we make the expected min and max chart areas visible to the user, and to the `color_calibrate.py` script.

Minor changes:
- Also adding `num_squares_w` and `num_squares_h` to `color_calibrate.py` to accept other chart formats
- Making size of morphological element used for dilation, closing, and small object removal dependent on `minAreaChart` and `maxAreaChart`

Reviewed By: fbriggs

Differential Revision: D4306491

fbshipit-source-id: 8c880c60a49c3b046b4b63143b434fdb5d263410