Bezier-based vignetting correction
Summary:
This diff improves the current Bezier-based vignetting correction to support dimension separability, and corrects a bug that was causing the vignetting function to stretch when the frame is not square.

Also, it makes the Bezier surface be centered around the center of the image, so the indices are easier to work with.

The default ISPs are also modified to reflect these changes.

Differential Revision: D4288642

fbshipit-source-id: 71cba9736fa4c1451ec07f6a3a6d233228a5b9c7