Bezier surface generation from data points
Summary:
This diff introduces `TestVignettingCalibration`, which takes as input a set of data points and generates the closest multi-channel, separable fourth order Bezier surface.

The main purpose of this is to generate surfaces for vignette removal, the points on the input file are supposed to be samples that model light falloff.

The input file is assumed to be a json file with the following format:

  [
    {
      "location" : [x1, y1],
      "rgbmedian" : [r1, g1, b1]
    },
    ...
    {
      "location" : [xN, yN],
      "rgbmedian" : [rN, gN, bN]
    }
   ]

When given a test image and ISP config file, the program also returns a modified ISP file with the generated Bezier control points, so it can be directly used.

NOTE: The input ISP file must not have a `vignetteRollOff` attribute, since the new format replaces it with `vignetteRollOffX` and `vignetteRollOffY`.

Reviewed By: tox-fb

Differential Revision: D4292856

fbshipit-source-id: e045f55f75596c477b0c62ef4bb7610d19092523