Color calibration: Lab + joint optimization
Summary:
This diff modifies the current color calibration to introduce two new approaches:
- Use Lab optimization for CCM computation, instead of RGB: Lab is perceptually uniform, which has nonlinear relations intended to mimic the human eye response.
- Jointly optimize for black level, white balance, and CCM: we can get closer to the ground truth by letting all the ISP parameters fluctuate. Advantage: less steps, so less prone to accumulated error.

A possible disadvantage of jointly optimizing for all parameters is that the black level is supposed to be a preset parameter by the sensor manufacturer, but we're letting the optimizer find the best value to minimize the overall cost. That's why we are not removing the existing black level detection feature, which let's you get the black level from a black hole on the image. Then we can lock that value on the optimization. Note that this is optional.

The optimization is basically the following:

We want to minimize

  sum_i=1^N ||LABgt_i - LAB(M * (s_i * RGBin_i - BL) / (1 - BL))||^2

where

N: number of color patches
LABgt_i: lab ground truth for i-th color patch
M: 3x3 matrix (can be interpreted as WB * CCM)
s_i: non-uniform illumination on i-th patch
RGBin_i: RGB value of color patch (input)
BL: black level

We assume that the color chart is not perfectly uniformly illuminated, so that each patch will be scaled by the amount of light it receives. However, we force all the illuminant scalars to sit on a smooth Bezier surface, to mimic real scenarios.

We solve the problem in Ceres.

Note that white balance and CCM are expressed as a single 3x3 matrix. This is because we want to constrain each component so that
- White balance just gains less sensitive channels to match more sensitive channel.
- CCM rows sum to 1, so that it does not introduce any extra gain, and the resulting RGB values don't fall out of gamut.

Here's how we do it:

  M = CCM * WB
  If GT is [1, 1, 1] and input is x = [x1, x2, x3], we want M * x = [a, a, a]
  (= keep grays gray)
  => x = M^-1 * [a, a, a]
  => WB = diag(1 / N(x)), where N(xi) = xi / max(x)
  (= scaling 2 channels to the most sensitive one)
  => WB * x = [b, b, b]
  CCM * WB * x = [a, a, a]
  => CCM * [b, b, b] = [a, a, a]
  => sum(CCM ith row) = a / b, for all i
  => CCM /= sum(CCM row) so each row sums to one

Other changes/notes:
- Ability to use accelerated ISP via Halide if the project is compiled with it.
- Renamed some flags to make them more understandable (e.g. `black_level_darkest` to `black_level_hole`, since we are actually finding a black hole on the image.)
- New color difference DeltaE lab errors: sqrt((L1-L2)^2 + (a1-a2)^2 + (b1-b2)^2)
- The clamping is directly added to the ISP config file, but not used for optimization.

Differential Revision: D4360994

fbshipit-source-id: bfd436d0b2018018b20563045ac1a376372a5a02