Script for feature extraction and matching, and geometric calibration
Summary:
Using COLMAP (https://colmap.github.io/) for feature extraction and matching. We make system calls to get a database of matches first, which we parse to a json file that `GeometricCalibration` can read.

Everything is on a python script, so that we can perform geometric calibration with a single call to `geometric_calibration.py`.

Differential Revision: D4479227

fbshipit-source-id: 20e7538d7aabd95ea1cbbf40a2e617c8eae26efc