Unpacker: non thread-safe serial number set
Summary:
This diff splits the set that holds serial numbers during unpacking to a separate set per binary file.

This is done because the sync functions that populate the serial number set are not thread safe, and sometimes there can be collisions when inserting elements to the set. This causes the set to not have some serial numbers, which causes directory renaming issues after unpacking (we rename directories from serial numbers to camera numbers, 0..N).

Minor change:
- Added log to show what serial number we're expecting. This saves us a lot of time when debugging.

Reviewed By: tox-fb

Differential Revision: D4880895

fbshipit-source-id: 3107c071e2c5319e3b77df1d0250aeca696a66a5