Add new sample data and instructions
Summary:
Data samples on GitHub are outdated and not compatible with the way we currently render.

We have captured, unpacked and rendered new sample data, and uploaded it to S3. Paths are specified on the top level README.

Also added instructions on how to generate an alternating left-right GIF to check stereo on a single frame.

Reviewed By: tox-fb

Differential Revision: D4979763

fbshipit-source-id: 21cceea80a9fa31b49745baba59f4ef395f82c45